+++ packages/eslint-plugin-wpcalypso/lib/configs/index.js
@@ -1,0 +1,7 @@
+const wpcalypso = import( './wpcalypso.js' );
+const react = import( './react.js' );
+
+module.exports = {
+	wpcalypso,
+	react,
+};